+++ packages/components/pie-modal/test/visual/pie-modal.spec.ts
@@ -1,4 +1,5 @@
 import { test } from '@sand4rt/experimental-ct-web';
+import { percyWidths } from '@justeattakeaway/pie-webc-testing/src/percy/breakpoints.ts';
 import percySnapshot from '@percy/playwright';
 import { positions } from '@/defs.ts';
 import { PieModal } from '@/index';
@@ -420,8 +421,8 @@ test.describe('Prop: `isFooterPinned`', () => {
             await mount(PieModal, {
                 props: {
                     heading: 'This is a modal heading',
-                    isOpen: true,
                     isFooterPinned,
+                    isOpen: true,
                     leadingAction: {
                         text: 'Confirm',
                         variant: 'primary',
@@ -458,6 +459,33 @@ test.describe('Prop: `isFooterPinned`', () => {
             });
 
             await percySnapshot(page, `Modal isFooterPinned: ${isFooterPinned}`);
+        });
+
+        (['medium', 'large'] as Array<ModalProps['size']>).forEach((size) => {
+            test(`when modal is fullscreen with size: ${size} and isFooterPinned: ${isFooterPinned}`, async ({ mount, page }) => {
+                await mount(PieModal, {
+                    props: {
+                        heading: 'This is a modal heading',
+                        isFooterPinned,
+                        isOpen: true,
+                        leadingAction: {
+                            text: 'Confirm',
+                            variant: 'primary',
+                            ariaLabel: 'Confirmation text',
+                        },
+                        size,
+                        isFullWidthBelowMid: size === 'medium',
+                    } as ModalProps,
+                    slots: {
+                        default: `<p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Deleniti fugit id exercitationem repellendus in magni
+                        quis obcaecati laboriosam est vero, perspiciatis ratione porro dolore repudiandae ea numquam! Ipsa, fugiat aut.Lorem ipsum dolor
+                        sit amet consectetur adipisicing elit. Deleniti fugit id exercitationem repellendus in magni quis obcaecati laboriosam est vero,
+                        perspiciatis ratione porro dolore repudiandae ea numquam! Ipsa, fugiat aut.</p>`,
+                    },
+                });
+
+                await percySnapshot(page, `Modal isFooterPinned: ${isFooterPinned}, fullscreen with size: ${size}`, percyWidths);
+            });
         });
     });
 });