+++ packages/components/pie-switch/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @justeattakeaway/pie-switch
 
+## 0.18.3
+
+### Patch Changes
+
+- Updated dependencies [[`e0516f18d`](https://github.com/justeattakeaway/pie/commit/e0516f18d6637c12a3efb5d692d3ca54e86f1a29)]:
+  - @justeattakeaway/pie-icons-webc@0.14.0
+
 ## 0.18.2
 
 ### Patch Changes